Split time iface from GPIO
Moved the time interface code into its own c/h files.

This will cause problems with drivers that used this interfce, but they'll just need to include this additional header
